The theme state (useState) and useEffect for initializing and updating the theme have been moved to ThemeContext
